test symlinks with multiple output situation
