+++ src/source/copy_dir.rs
@@ -35,21 +35,29 @@ impl Default for CopyOptions {
 }
 
 /// Cross platform way of creating a symlink
+/// Creates a symlink from `link` to `original`
+/// The file that is newly created is the `link` file
 pub(crate) fn create_symlink(
-    from: impl AsRef<Path>,
-    to: impl AsRef<Path>,
+    original: impl AsRef<Path>,
+    link: impl AsRef<Path>,
 ) -> Result<(), SourceError> {
-    let from = from.as_ref();
-    let to = to.as_ref();
+    let original = original.as_ref();
+    let link = link.as_ref();
 
-    if from.exists() {
-        fs_err::remove_file(from)?;
+    if link.exists() {
+        fs_err::remove_file(link)?;
     }
-    println!("Creating symlink from {:?} to {:?}", from, to);
+
     #[cfg(unix)]
-    fs_err::os::unix::fs::symlink(to, from)?;
+    fs_err::os::unix::fs::symlink(original, link)?;
     #[cfg(windows)]
-    std::os::windows::fs::symlink_file(to, from)?;
+    {
+        if original.is_dir() {
+            std::os::windows::fs::symlink_dir(original, link)?;
+        } else {
+            std::os::windows::fs::symlink_file(original, link)?;
+        }
+    }
 
     Ok(())
 }
@@ -77,7 +85,7 @@ pub(crate) fn copy_file(
         // if file is a symlink, copy it as a symlink
         if path.is_symlink() {
             let link_target = fs_err::read_link(path)?;
-            create_symlink(dest_path, link_target)?;
+            create_symlink(link_target, dest_path)?;
             Ok(())
         } else {
             if dest_path.exists() {
@@ -509,7 +517,7 @@ mod test {
             .use_gitignore(false)
             .run()
             .unwrap();
-        println!("{:?}", copy_dir.copied_paths());
+
         assert_eq!(copy_dir.copied_paths().len(), 2);
         let expected = [
             dest_dir_3.join("test_dir/test.md"),
